RUST-1366 Create serde_util module
