+++ src/test/spec/unified_runner/test_file.rs
@@ -28,7 +28,8 @@ use crate::{
         SelectionCriteria,
         WriteConcern,
     },
-    test::{serde_helpers::deserialize_nonempty_vec, Serverless, TestClient, DEFAULT_URI},
+    serde_util,
+    test::{Serverless, TestClient, DEFAULT_URI},
 };
 
 #[derive(Debug, Deserialize)]
@@ -37,11 +38,11 @@ pub(crate) struct TestFile {
     pub(crate) description: String,
     #[serde(deserialize_with = "deserialize_schema_version")]
     pub(crate) schema_version: Version,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) run_on_requirements: Option<Vec<RunOnRequirement>>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) create_entities: Option<Vec<TestFileEntity>>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) initial_data: Option<Vec<CollectionData>>,
     pub(crate) tests: Vec<TestCase>,
     // We don't need to use this field, but it needs to be included during deserialization so that
@@ -71,7 +72,7 @@ where
 pub(crate) struct RunOnRequirement {
     min_server_version: Option<String>,
     max_server_version: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     topologies: Option<Vec<Topology>>,
     server_parameters: Option<Document>,
     serverless: Option<Serverless>,
@@ -178,15 +179,15 @@ pub(crate) struct Client {
     pub(crate) id: String,
     pub(crate) uri_options: Option<Document>,
     pub(crate) use_multiple_mongoses: Option<bool>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) observe_events: Option<Vec<ObserveEvent>>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) ignore_command_monitoring_events: Option<Vec<String>>,
     #[serde(default)]
     pub(crate) observe_sensitive_commands: Option<bool>,
     #[serde(default, deserialize_with = "deserialize_server_api_test_format")]
     pub(crate) server_api: Option<ServerApi>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) store_events_as_entities: Option<Vec<StoreEventsAsEntity>>,
     #[cfg(feature = "tracing-unstable")]
     #[serde(default, deserialize_with = "deserialize_tracing_level_map")]
@@ -392,15 +393,15 @@ pub(crate) struct CollectionData {
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
 pub(crate) struct TestCase {
     pub(crate) description: String,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) run_on_requirements: Option<Vec<RunOnRequirement>>,
     pub(crate) skip_reason: Option<String>,
     pub(crate) operations: Vec<Operation>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) expect_events: Option<Vec<ExpectedEvents>>,
     #[cfg(feature = "tracing-unstable")]
     pub(crate) expect_log_messages: Option<Vec<ExpectedMessages>>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) outcome: Option<Vec<CollectionData>>,
 }
 
@@ -460,9 +461,9 @@ pub(crate) struct ExpectError {
     pub(crate) error_contains: Option<String>,
     pub(crate) error_code: Option<i32>,
     pub(crate) error_code_name: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) error_labels_contain: Option<Vec<String>>,
-    #[serde(default, deserialize_with = "deserialize_nonempty_vec")]
+    #[serde(default, deserialize_with = "serde_util::deserialize_nonempty_vec")]
     pub(crate) error_labels_omit: Option<Vec<String>>,
     pub(crate) expect_result: Option<Bson>,
 }
